feat: move client to specified dir and add mapping profile for proto
